Complete code for Online photography contest
It is a website which allows authenticated users to upload their photos
for the contest and allows anyone to view the photos and viewers  can
vote. An image optimisation algorithm is used which crops the center
part of the image and displays it as a thumbnail and for each photo an
unique URL,Facebook like and forum is dynamically generated.